Move displaying game score to another class
For separating responsibility between classes
