+++ tennis_kata_referee/src/main/java/TennisGame.java
@@ -1,40 +1,20 @@
-import java.util.ArrayList;
-import java.util.Arrays;
-
 class TennisGame {
 
     private int federerScore;
     private int nadalScore;
 
-    private ArrayList<String> gameScoreHeader;
-    private ArrayList<String> gameScoreFedererRow;
-    private ArrayList<String> gameScoreNadalRow;
-    private ArrayList<String> gameScoreWinnerRow;
+    private TennisGameScore tennisGameScore;
 
     TennisGame() {
         this.federerScore = 0;
         this.nadalScore = 0;
 
-        this.gameScoreHeader = new ArrayList<>();
-        this.gameScoreFedererRow = new ArrayList<>();
-        this.gameScoreNadalRow = new ArrayList<>();
-        this.gameScoreWinnerRow = new ArrayList<>();
+        this.tennisGameScore = new TennisGameScore();
     }
 
     void play() {
         System.out.println("Game score");
-
-        this.gameScoreHeader.add("");
-        this.gameScoreHeader.add("Start the game");
-
-        this.gameScoreFedererRow.add("Federer");
-        this.gameScoreFedererRow.add("0");
-
-        this.gameScoreNadalRow.add("Nadal");
-        this.gameScoreNadalRow.add("0");
-
-        this.gameScoreWinnerRow.add("");
-        this.gameScoreWinnerRow.add("");
+        this.tennisGameScore.init();
 
         int playResult;
         String winner;
@@ -50,37 +30,16 @@ class TennisGame {
                 winner = "Nadal";
             }
 
-            this.gameScoreHeader.add(winner + " wins 1 point");
+            this.tennisGameScore.addWinnerPoint(winner);
 
             if (isTennisGameOver()) {
-                this.gameScoreFedererRow.add("0");
-                this.gameScoreNadalRow.add("0");
-                this.gameScoreWinnerRow.add(winner + " wins the game");
+                this.tennisGameScore.addGameOver(winner);
             } else {
-                this.gameScoreFedererRow.add(convertToTennisScore(this.federerScore));
-                this.gameScoreNadalRow.add(convertToTennisScore(this.nadalScore));
-                this.gameScoreWinnerRow.add("");
+                this.tennisGameScore.addPoint(this.federerScore, this.nadalScore);
             }
-
         }
 
-        displayGameScore();
-    }
-
-    private void displayGameScore() {
-        TennisGameTable tennisGameTable = new TennisGameTable();
-
-        tennisGameTable.setShowVerticalLines(true);
-        tennisGameTable.setHeaders(toArrayString(this.gameScoreHeader));
-        tennisGameTable.addRow(toArrayString(this.gameScoreFedererRow));
-        tennisGameTable.addRow(toArrayString(this.gameScoreNadalRow));
-        tennisGameTable.addRow(toArrayString(this.gameScoreWinnerRow));
-
-        tennisGameTable.print();
-    }
-
-    private String[] toArrayString(ArrayList<String> arrayList) {
-        return Arrays.copyOf(arrayList.toArray(), arrayList.toArray().length, String[].class);
+        this.tennisGameScore.displayGameScore();
     }
 
     private boolean isTennisGameOver() {
@@ -93,16 +52,6 @@ class TennisGame {
 
     private boolean isFedererWinPoint(int randResult) {
         return randResult % 2 == 0;
-    }
-
-    private String convertToTennisScore(int score) {
-        switch(score) {
-            case 1: return "15";
-            case 2: return "30";
-            case 3: return "40";
-        }
-
-        return "0";
     }
 
 }
